requirements.txt
alembic==0.8.10
apache-airflow==1.8.2
appdirs==1.4.3
attrs==17.2.0
bleach==1.5.0
bokeh==0.12.5
certifi==2016.2.28
click==6.7
contextlib2==0.5.5
croniter==0.3.19
cycler==0.10.0
decorator==4.1.2
dill==0.2.7.1
dlib==19.5.1
docopt==0.6.2
docutils==0.14
eli5==0.6.2
entrypoints==0.2.2
enum34==1.1.6
Flask==0.12
Flask-Admin==1.4.1
Flask-Cache==0.13.1
Flask-Cors==3.0.2
Flask-Login==0.2.11
flask-swagger==0.2.13
Flask-WTF==0.14
funcsigs==1.0.0
future==0.16.0
gevent==1.2.1
git-review==1.25.0
gitdb==0.6.4
gitdb2==2.0.2
GitPython==2.0.8
glove-python==0.1.0
graphviz==0.7.1
greenlet==0.4.12
gunicorn==19.3.0
h5py==2.7.0
html5lib==0.9999999
humanize==0.5.1
ipykernel==4.6.1
ipython==6.0.0
ipython-genutils==0.2.0
ipywidgets==5.0.0
itsdangerous==0.24
jedi==0.10.2
Jinja2==2.9.6
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.0.1
jupyter-console==5.1.0
jupyter-core==4.3.0
Keras==2.0.8
keras-vis==0.2.1
kitchen==1.2.4
lime==0.1.1.20
lockfile==0.12.2
lxml==3.8.0
Mako==1.0.7
Markdown==2.6.9
MarkupSafe==1.0
matplotlib==2.0.2
mistune==0.7.4
mock==2.0.0
more-itertools==2.2
nbconvert==5.1.1
nbformat==4.3.0
neptune-cli==2.1.4
networkx==1.11
notebook==5.0.0
numpy==1.13.3
oauthlib==2.0.4
olefile==0.44
opencv-python==3.2.0.7
ordereddict==1.1
packaging==16.8
pandas==0.20.3
pandocfilters==1.4.1
pathlib2==2.3.0
pbr==3.1.1
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.3.0
prompt-toolkit==1.0.14
protobuf==3.4.0
psutil==4.3.0
ptyprocess==0.5.1
Pygments==2.2.0
PyJWT==1.5.2
pykwalify==1.5.2
pymongo==3.5.1
pyparsing==2.2.0
python-daemon==2.1.2
python-dateutil==2.4.2
python-editor==1.0.3
python-nvd3==0.14.2
python-slugify==1.1.4
pytz==2017.2
PyWavelets==0.5.2
PyYAML==3.11
pyzmq==16.0.2
qtconsole==4.3.0
quiver-engine==0.1.4.1.4
raven==6.1.0
requests==2.11.1
requests-oauthlib==0.8.0
scikit-image==0.13.0
scikit-learn==0.19.0
scipy==0.19.1
seaborn==0.7.1
setproctitle==1.1.10
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
sklearn==0.0
smmap==0.9.0
smmap2==2.0.3
SQLAlchemy==1.1.14
tabulate==0.7.7
tee==0.0.2
tensorflow-gpu==1.1.0
tensorflow-tensorboard==0.1.6
terminado==0.6
terminaltables==2.1.0
testpath==0.3
Theano==0.9.0
thrift==0.9.3
tornado==4.2
tqdm==4.11.2
traitlets==4.3.2
typing==3.6.1
ujson==1.35
Unidecode==0.4.21
urllib3==1.16
voluptuous==0.9.3
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.35.0
Werkzeug==0.12.2
widgetsnbextension==2.0.0
WTForms==2.1
zope.deprecation==4.3.0
